refactor(Alarms): created 'components' folder and moved into 'MonthToggleButton' and 'MonthTogglePanel'
